pyproject.toml
[tool.poetry]
name = "nuggets"
version = "0.1.0"
description = ""
authors = ["Samir Azzouz <s.azzouz97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
djangorestframework = "^3.12.4"
graphene-django = "^2.15.0"
django-taggit = "^1.5.1"
graphene = "^2.1.9"
django-filter = "^2.4.0"
django-tailwind = "^2.2.0"
Django = "^3.2.6"
black = "^21.7b0"
Markdown = "^3.3.4"
django-crispy-forms = "^1.12.0"
django-debug-toolbar = "^3.2.1"
Pillow = "^8.3.1"
django-rest-auth = "^0.9.5"
django-allauth = "^0.45.0"
coreapi = "^2.3.3"
PyYAML = "^5.4.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^4.7.2"
django-model-utils = "^4.1.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
